Refactor chat-content component and add custom hooks for estimated waiting time and name/title visibility
