+++ src/components/chat-content/chat-content.tsx
@@ -3,21 +3,16 @@ import { useEffect, useRef } from 'react';
 import { OverlayScrollbarsComponent } from 'overlayscrollbars-react';
 import ChatMessage from '../chat-message/chat-message';
 import useChatSelector from '../../hooks/use-chat-selector';
-import { useAppDispatch } from '../../store';
-import { getEstimatedWaitingTime, getNameVisibility, getTitleVisibility, setEstimatedWaitingTimeToZero } from '../../slices/chat-slice';
-import 'overlayscrollbars/css/OverlayScrollbars.css';
-import './os-custom-theme.scss';
 import styles from './chat-content.module.scss';
 import WaitingTimeNotification from '../waiting-time-notification/waiting-time-notification';
+import useEstimatedWaitingTime from '../../hooks/use-estimated-waiting-time';
+import useNameAndTitleVisibility from '../../hooks/use-name-title-visibility';
+import 'overlayscrollbars/css/OverlayScrollbars.css';
+import './os-custom-theme.scss';
 
 const ChatContent = (): JSX.Element => {
   const OSref = useRef<OverlayScrollbarsComponent>(null);
-  const { 
-    messages, 
-    estimatedWaiting, 
-    customerSupportId, 
-  } = useChatSelector();
-  const dispatch = useAppDispatch();
+  const { messages } = useChatSelector();
 
   useEffect(() => {
     if (OSref.current) {
@@ -26,24 +21,8 @@ const ChatContent = (): JSX.Element => {
     }
   }, [messages]);
 
-  useEffect(() => {
-    const interval = setInterval(() => {
-      if(!customerSupportId) {
-        dispatch(getEstimatedWaitingTime());
-      } else {
-        dispatch(setEstimatedWaitingTimeToZero());
-      }
-    }, 2000);
-
-    return () => clearInterval(interval);
-  }, [customerSupportId]);
-
-  useEffect(() => {
-    if (!customerSupportId) {
-      dispatch(getNameVisibility());
-      dispatch(getTitleVisibility());
-    }
-  }, [customerSupportId]);
+  useEstimatedWaitingTime();
+  useNameAndTitleVisibility();
 
   return (
     <AnimatePresence initial={false}>
